feature: AP-7-Created Container with main layout, added @imports added logout to avatar
