+++ package.json
@@ -17,6 +17,7 @@
     "@mui/icons-material": "^6.3.1",
     "@mui/material": "^6.3.1",
     "@reduxjs/toolkit": "^2.5.0",
+    "@types/node": "^22.10.5",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-hook-form": "^7.54.2",
