+++ src/components/organisms/NavigationMenu/NavigationMenu.tsx
@@ -1,19 +1,33 @@
 import type { ReactNode } from "react";
 import { APDrawer, APNavBar, useAPDrawer } from "../../molecules";
 import { navigationPanelItems } from "./constants/navigationPanelItems";
-import { userMenuItems } from "./constants/userMenuItems";
+import { useDispatch } from "react-redux";
+import { useNavigate } from "react-router-dom";
+import { logout } from "@/services/auth";
+import { createUserMenuItems } from "./helpers";
 
 interface INavigationMenu {
   children?: ReactNode;
 }
 
 export const NavigationMenu = ({ children }: INavigationMenu) => {
+  const dispatch = useDispatch();
+  const navigate = useNavigate();
+
+  const handleLogout = () => {
+    dispatch(logout());
+    navigate("/login");
+  };
+
   const {
     mobileOpen,
     handleDrawerClose,
     handleDrawerTransitionEnd,
     handleDrawerToggle,
   } = useAPDrawer();
+
+  const userMenuItems = createUserMenuItems(handleLogout);
+
   return (
     <>
       <APDrawer
