+++ src/services/auth/index.ts
@@ -1,3 +1,4 @@
+export { logout } from "./authState";
 export { selectAccessToken } from "./constants/authSelectors";
 export { selectRefreshToken } from "./constants/authSelectors";
 export { selectAuthUser } from "./constants/authSelectors";
